Only attempt to copy built apk when it exists
This is a temporary workaround for an issue that occurs when building
an app that uses Crosswalk.

We usually attempt to copy android-release-unsigned.apk to the root
build dir after a successful release build, but because Crosswalk
depends on native code, the resulting apk is architecture-specific
(android-armv7-release-unsigned.apk and
android-x86-release-unsigned.apk).

For now, I’m just adding a check so we don’t throw ENOENT. Built apks
can be found as usual in android/project/build/outputs/apk, they’re
just not copied to the root build dir.

+++ tools/cli/commands.js
@@ -984,8 +984,10 @@ https://github.com/meteor/meteor/wiki/How-to-submit-your-iOS-app-to-App-Store
               const apkPath = files.pathJoin(buildPath, 'build/outputs/apk',
                 options.debug ? 'android-debug.apk' : 'android-release-unsigned.apk')
 
+              if (files.exists(apkPath)) {
               files.copyFile(apkPath, files.pathJoin(platformOutputPath,
                 options.debug ? 'debug.apk' : 'release-unsigned.apk'));
+              }
 
               files.writeFile(
                 files.pathJoin(platformOutputPath, 'README'),
